parser/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
pymupdf = "^1.24.14"
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"


[tool.poetry.group.dev.dependencies]
ruff = "0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
